+++ tests/test_modeling.py
@@ -158,6 +158,38 @@ def test_gaussian_fit_frozen_and_bound():
     plt.legend()
 
 
+def test_gaussian_fit_iterative_hist():
+    """Test an iterative fit for a Gaussian histogram.
+    """
+    plt.figure(inspect.currentframe().f_code.co_name)
+    model = Gaussian()
+    print(model)
+    TEST_HISTOGRAM.plot()
+    model.fit_iterative(TEST_HISTOGRAM, num_iterations=3, num_sigma_left=3., num_sigma_right=3.)
+    model.plot(fit_output=True)
+    plt.legend()
+    assert model.mu.compatible_with(0., NUM_SIGMA)
+    assert model.sigma.compatible_with(1., NUM_SIGMA)
+    assert model.status.pvalue > 0.001
+
+
+def test_gaussian_fit_iterative_scatter():
+    """Test an iterative fit for a Gaussian scatter plot.
+    """
+    plt.figure(inspect.currentframe().f_code.co_name)
+    model = Gaussian(xlabel="x [a.u.]", ylabel="y [a.u.]")
+    print(model)
+    sigma = 0.01
+    xdata, ydata = model.random_fit_dataset(sigma, seed=313)
+    plt.errorbar(xdata, ydata, sigma, fmt="o", label="Random data")
+    model.fit_iterative(xdata, ydata, sigma=sigma, num_iterations=3)
+    model.plot(fit_output=True)
+    plt.legend()
+    assert model.mu.compatible_with(0., NUM_SIGMA)
+    assert model.sigma.compatible_with(1., NUM_SIGMA)
+    assert model.status.pvalue > 0.001
+
+
 def test_sum_gauss_line():
     """Test the sum of of two models.
     """